move custom output formatter into separate class/file
